Fixed package names to conform to Maven file structure
