+++ src/main/java/com/teragrep/jos_01/procfs/status/Status.java
@@ -43,44 +43,21 @@
  * Teragrep, the applicable Commercial License may apply to this file if you as
  * a licensee so wish it.
  */
-package src.main.java.procfs;
+package com.teragrep.jos_01.procfs.status;
 
-import src.main.java.procfs.status.GenericStatus;
-import src.main.java.procfs.status.OSStat;
-
-import java.io.File;
+import java.time.LocalDateTime;
 import java.util.ArrayList;
+import java.util.Map;
 
-public class OS {
+public interface Status {
 
-    private final File procDirectory;
+    void printStatistics();
 
-    public OS() {
-        procDirectory = new File("/proc");
-    }
+    void printTimestamp();
 
-    private ArrayList<String> readProcFile(String procFileName) {
-        ProcFile procFile = new ProcFile(procDirectory, procFileName);
-        ArrayList<String> rows = procFile.readFile();
-        return rows;
-    }
+    Map<String, String> statistics();
 
-    public OSStat stat() {
-        ArrayList<String> rows = readProcFile("stat");
-        return new OSStat(rows);
-    }
+    ArrayList<String> rows();
 
-    // Estimates page size in kB. TODO: implement specific status object for memInfo and vmStat to make getting these fields prettier
-    float pageSize() {
-        String field = proc("meminfo").rows().get(21).split(":")[1].split("kB")[0].trim();
-        float mapped = Long.parseLong(field);
-        float nr_mapped = Long.parseLong(proc("vmstat").rows().get(35).split(" ")[1]);
-        float pageSize = mapped / nr_mapped;
-        return pageSize;
-    }
-
-    public GenericStatus proc(String procFileName) {
-        ArrayList<String> rows = readProcFile(procFileName);
-        return new GenericStatus(rows);
-    }
+    LocalDateTime timestamp();
 }
